requirements.txt
aiohttp
aiosignal
annotated-types
anyio
appnope
asttokens
asyncio
attrs
certifi
cffi
charset-normalizer
comm
dataclasses-json
debugpy
decorator
deepgram-sdk
distro
executing
frozenlist
gevent
greenlet
groq
h11
httpcore
httpx
idna
ipykernel
ipython
jedi
jsonpatch
jsonpointer
jupyter_client
jupyter_core
langchain
langchain-community
langchain-core
langchain-groq
langchain-openai
langsmith
marshmallow
matplotlib-inline
multidict
mypy-extensions
nest-asyncio
numpy
openai
orjson
packaging
parso
pexpect
platformdirs
prompt-toolkit
psutil
ptyprocess
pure-eval
PyAudio
pycparser
pydantic
pydantic_core
pygame
Pygments
python-dateutil
python-dotenv
PyYAML
pyzmq
regex
requests
six
sniffio
sounddevice
SQLAlchemy
stack-data
tenacity
tiktoken
tornado
tqdm
traitlets
typing-inspect
typing_extensions
urllib3
verboselogs
wcwidth
websocket
websocket-client
websockets
yarl
zope.event
zope.interface
twilio
flask
flask-cors
schedule
